Moved the rewarding methods from `VM` to `VMState` and added `VMState.finalize_block(block)`
