+++ evm/vm/forks/byzantium/vm_state.py
@@ -1,11 +1,18 @@
+from evm.constants import (
+    MAX_UNCLE_DEPTH,
+)
 from evm.rlp.receipts import (
     Receipt,
+)
+from evm.validation import (
+    validate_lte,
 )
 from evm.vm.forks.frontier.vm_state import _make_frontier_receipt
 from evm.vm.forks.spurious_dragon.vm_state import SpuriousDragonVMState
 
 from .computation import ByzantiumComputation
 from .constants import (
+    EIP649_BLOCK_REWARD,
     EIP658_TRANSACTION_STATUS_CODE_FAILURE,
     EIP658_TRANSACTION_STATUS_CODE_SUCCESS,
 )
@@ -28,3 +35,13 @@ class ByzantiumVMState(SpuriousDragonVMState):
             logs=old_receipt.logs,
         )
         return receipt
+
+    @staticmethod
+    def get_block_reward():
+        return EIP649_BLOCK_REWARD
+
+    @staticmethod
+    def get_uncle_reward(block_number, uncle):
+        validate_lte(uncle.block_number, MAX_UNCLE_DEPTH)
+        block_number_delta = block_number - uncle.block_number
+        return (8 - block_number_delta) * EIP649_BLOCK_REWARD // 8
