+++ evm/vm/forks/frontier/__init__.py
@@ -1,10 +1,6 @@
 from __future__ import absolute_import
 
 from evm import VM
-from evm.constants import (
-    BLOCK_REWARD,
-    UNCLE_DEPTH_PENALTY_FACTOR,
-)
 
 from .blocks import FrontierBlock
 from .vm_state import FrontierVMState
@@ -13,20 +9,6 @@ from .headers import (
     compute_frontier_difficulty,
     configure_frontier_header,
 )
-
-
-def _frontier_get_block_reward():
-    return BLOCK_REWARD
-
-
-def _frontier_get_uncle_reward(block_number, uncle):
-    return BLOCK_REWARD * (
-        UNCLE_DEPTH_PENALTY_FACTOR + uncle.block_number - block_number
-    ) // UNCLE_DEPTH_PENALTY_FACTOR
-
-
-def _frontier_get_nephew_reward(cls):
-    return cls.get_block_reward() // 32
 
 
 FrontierVM = VM.configure(
@@ -38,9 +20,6 @@ FrontierVM = VM.configure(
     create_header_from_parent=staticmethod(create_frontier_header_from_parent),
     compute_difficulty=staticmethod(compute_frontier_difficulty),
     configure_header=configure_frontier_header,
-    get_block_reward=staticmethod(_frontier_get_block_reward),
-    get_uncle_reward=staticmethod(_frontier_get_uncle_reward),
-    get_nephew_reward=_frontier_get_nephew_reward,
     # mode
     _is_stateless=True,
 )
